requirements.txt
azure-storage-blob
chromadb==0.4.8
langchain==0.0.281
openai==0.28.0
pandas
pypdf
streamlit
tiktoken
